# api/openapi.yaml
openapi: 3.0.3
info:
  title: TalkFlow API
  description: API for TalkFlow, a smart video/audio call application with AI-powered notes.
  version: 1.0.0
  contact:
    name: TalkFlow Support
    email: support@talkflow.example.com
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: https://api.talkflow.example.com/api/v1
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
...
components:
  schemas:
...
    Meeting:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
          nullable: true
        hostId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/MeetingStatus'
        notesEnabled:
          type: boolean
        roomCode:
          type: string
        startedAt:
          type: string
          format: date-time
          nullable: true
        endedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    MeetingStatus:
      type: string
      enum: [SCHEDULED, ACTIVE, ENDED, CANCELLED]
...
paths:
...
  /meetings/{meetingId}:
    get:
      tags:
        - Meetings
      summary: Get a specific meeting
      parameters:
        - name: meetingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: The requested meeting
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Meeting'
    patch:
      tags:
        - Meetings
      summary: Update a specific meeting
      parameters:
        - name: meetingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                  nullable: true
                notesEnabled:
                  type: boolean
      responses:
        '200':
          description: Meeting updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Meeting'
    delete:
      tags:
        - Meetings
      summary: Delete a specific meeting
      parameters:
        - name: meetingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Meeting deleted successfully
...